Bug 1533694 - prepare next task user before using current task user
