+++ plat_all-unix-style.go
@@ -8,12 +8,11 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"runtime"
-	"strconv"
+	sysruntime "runtime"
 	"strings"
-	"time"
 
 	"github.com/taskcluster/generic-worker/process"
+	"github.com/taskcluster/generic-worker/runtime"
 	"github.com/taskcluster/shell"
 )
 
@@ -32,15 +31,9 @@ func (pd *PlatformData) ReleaseResources() error {
 	return nil
 }
 
-type OSUser struct {
-	HomeDir  string
-	Name     string
-	Password string
-}
-
 type TaskContext struct {
 	TaskDir string
-	User    *OSUser
+	User    *runtime.OSUser
 }
 
 func immediateShutdown(cause string) {
@@ -94,10 +87,11 @@ func (task *TaskRun) generateCommand(index int) error {
 }
 
 func purgeOldTasks() error {
-	if config.CleanUpTaskDirs {
-		deleteTaskDirs()
+	if !config.CleanUpTaskDirs {
+		log.Printf("WARNING: Not purging previous task directories/users since config setting cleanUpTaskDirs is false")
+		return nil
 	}
-	return nil
+	return deleteTaskDirs()
 }
 
 func install(arguments map[string]interface{}) (err error) {
@@ -160,7 +154,7 @@ func (task *TaskRun) formatCommand(index int) string {
 }
 
 func prepareTaskUser(username string) bool {
-	taskContext.User = &OSUser{
+	taskContext.User = &runtime.OSUser{
 		Name: username,
 	}
 	err := os.MkdirAll(taskContext.TaskDir, 0777)
@@ -170,8 +164,8 @@ func prepareTaskUser(username string) bool {
 	return false
 }
 
-func deleteTaskDir(path string) error {
-	log.Print("Removing task directory '" + path + "'...")
+func deleteDir(path string) error {
+	log.Print("Removing directory '" + path + "'...")
 	err := os.RemoveAll(path)
 	if err != nil {
 		log.Print("WARNING: could not delete directory '" + path + "'")
@@ -187,22 +181,27 @@ func defaultTasksDir() string {
 	return filepath.Dir(os.Getenv("HOME"))
 }
 
-// N/A for unix - just a windows thing
-func AutoLogonCredentials() (string, string) {
-	return "", ""
-}
-
-func chooseTaskDirName() string {
-	return "task_" + strconv.Itoa(int(time.Now().Unix()))
-}
-
-func unsetAutoLogon() {
-}
-
-func deleteTaskDirs() {
-	removeTaskDirs(config.TasksDir)
+func deleteTaskDirs() error {
+	currentTaskUser := taskContext.User.Name
+	taskDirs, err := taskDirsIn(config.TasksDir)
+	if err != nil {
+		return err
+	}
+	for _, taskDir := range taskDirs {
+		if filepath.Base(taskDir) != currentTaskUser {
+			err = deleteDir(taskDir)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
 }
 
 func GrantSIDFullControlOfInteractiveWindowsStationAndDesktop(sid string) (err error) {
-	return fmt.Errorf("Cannot grant %v full control of interactive windows station and desktop; platform %v does not have such entities", sid, runtime.GOOS)
+	return fmt.Errorf("Cannot grant %v full control of interactive windows station and desktop; platform %v does not have such entities", sid, sysruntime.GOOS)
+}
+
+func rebootBetweenTasks() bool {
+	return false
 }
