+++ plat_darwin.go
@@ -2,14 +2,13 @@ package main
 
 import (
 	"io/ioutil"
-	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strconv"
 	"time"
 
 	"github.com/dchest/uniuri"
+	"github.com/taskcluster/generic-worker/runtime"
 )
 
 func createNewTaskContext() error {
@@ -19,12 +18,12 @@ func createNewTaskContext() error {
 	password := generatePassword()
 	taskContext = &TaskContext{
 		TaskDir: "/Users/" + userName,
-		User: &OSUser{
+		User: &runtime.OSUser{
 			Name:     userName,
 			Password: password,
 		},
 	}
-	err := taskContext.User.createNewOSUser()
+	err := taskContext.User.Create(false)
 	if err != nil {
 		return err
 	}
@@ -34,35 +33,6 @@ func createNewTaskContext() error {
 		return err
 	}
 	return os.MkdirAll(filepath.Join(taskContext.TaskDir, filepath.Dir(logPath)), 0777)
-}
-
-func (user *OSUser) createNewOSUser() error {
-
-	createUserScript := `
-		set -eu
-		username="${0}"
-		homedir="${1}"
-		fullname="${2}"
-		password="${3}"
-		echo "Creating user '${username}' with home directory '${homedir}' and password '${password}'..."
-		maxid=$(dscl . -list '/Users' 'UniqueID' | awk '{print $2}' | sort -ug | tail -1)
-		newid=$((maxid+1))
-		dscl . -create "/Users/${username}"
-		dscl . -create "/Users/${username}" 'UserShell' '/bin/bash'
-		dscl . -create "/Users/${username}" 'RealName' "${fullname}"
-		dscl . -create "/Users/${username}" 'UniqueID' "${newid}"
-		dscl . -passwd "/Users/${username}" "${password}" 
-		staff="$(dscl . -read /Groups/staff | awk '($1 == "PrimaryGroupID:") { print $2 }')"
-		dscl . -create "/Users/${username}" 'PrimaryGroupID' "${staff}"
-		dscl . -create "/Users/${username}" 'NFSTaskDirectory' "${homedir}"
-		cp -R '/System/Library/User Template/English.lproj' "${homedir}"
-		chown -R "${username}:staff" "${homedir}"
-		echo "User '${username}' created."
-	`
-
-	out, err := exec.Command("sudo", "/bin/bash", "-c", createUserScript, user.Name, user.HomeDir, user.Name+" User", user.Password).Output()
-	log.Print(string(out))
-	return err
 }
 
 // Uses [A-Za-z0-9] characters (default set) to avoid strange escaping problems
