+++ runtime/runtime_all-unix-style.go
@@ -1,0 +1,9 @@
+// +build !windows
+
+package runtime
+
+type OSUser struct {
+	HomeDir  string
+	Name     string
+	Password string
+}
